requirements.txt
# requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
gunicorn==20.1.0
Flask-WTF==1.0.1
python-keycloak==3.1.0
email_validator==1.3.0
requests==2.26.0